Lua: Add preloading of the ported V8 string primitives
These are required by the path library, but they don't need to be exposed to the global environment. I'm sure a better solution exists, but until I can think of one simply making them available as preloaded builtins seems best.

This way they can be required, but won't pollute the global namespace when they aren't needed (which is most of the time).