CI: Remove workflows for irrelevant Luvi variants
I've no need for any of these and I don't want to spend more time debugging build failures that shouldn't happen, so I'll just remove it (for now).

If needed, people it's still possible to use the official Luvi versions, but I'd rather only support the version I'm actually using and save myself a lot of headaches this way.