Lua: Always include epo packages in compiled bundles
Since this is technically a "hidden" folder, Luvi will exclude it when compiling bundles into a self-contained executable. The result is that packages installed in it won't be recognized.

Since it's hidden by design, and the dot marks it a special directory, whitelisting it seems like the easiest way to circumvent the limitations in a way that's both extensible and backwards-compatible.