Lua: Defer loading of primitives to ensure luvi is available
Without this, the virtual file system primitive can't access ``luvi.bundle`` since primitives are loaded during the bundle initialization process, but the initial ``require`` happened sooner (when ``luvi`` was not preloaded).

This doesn't change any functionality, since the bundle initialization is the only place actually using extensions and primitives (to embed them in the bundle).