Test: Fix missing assertion error bundle test for import app
Since the bundle won't include files from the parent directory, the assertion helper isn't added to it and this causes the respective test to fail.

A better design could likely solve this, but for now I don't care enough to bother. Assertions should likely be streamlined anyway, but that's a separate issue.