CI: Do not trigger duplicate workflow runs on PRs
Currently, a commit to an open PR will trigger two runs (push on the remote branch, and one for the PR itself). This change will eliminate the former if it's not a commit to the main branch itself. PRs also won't trigger a run unless they're targeting the main branch/