Lua: Fix "yield across C-call boundary" error source in import builtin
This is unfortunately very difficult to test without pulling in a ton of boilerplate code, as it requires an elaborate setup of user script wrapped in a  coroutine into asynchronous libuv callbacks into coroutine yield in the user thread to reproduce.

For the time being, relying on evo's test suite to discover issues with coroutines inside imports seems acceptable, even if it's not great. Since the functionality is identical, no other use cases should be affected.