Lua: Load imported modules from disk over the VFS
If trying to load a file that also exists in the runtime's bundle itself, the file will never be loaded from disk. This is somewhat unintuitive, even if it was a deliberate design decision at the time.

Since it's much easier to rename/remove local files than those bundled with an application (inside the executable), it might make more sense to prefer loading everything from disk and only fall back to looking into the zipped bundle itself when that has already failed.