Runtime: Introduce an error when attempting to combine bundles
These cause more headaches than they are worth, and it's doubtful anyone would really need this feature.

In the long term, the CLI interface should be streamlined and work similarly to the NodeJS runtime, which makes obsolete the code that runs multiple "layered" bundles. As the code quality is also very low, I'm removing it to get rid of the maintenance burden.