Runtime: Simplify the CLI handling of bundle paths
Since we can now assume that only a single application or script file will be loaded, it makes no sense to store it in an array.

This will make it easier to later transition to a different interface, where the first argument will be considered a script file, module directory, or zip app path to load, the following arguments will be the parameters passed (if not a valid flag), and the ugly -- hack can finally be removed.