Cleanup: Remove the nonstandard global args export
This is redundant and therefore pointless, because the arguments are already being passed as varargs when loading the script. Since it's also nonstandard, having this seems like a bad idea.

An argument could be made for exposing the args via the global arg (as in the standard Lua interpreter), but that's a separate issue.