Tests: Port busted unit tests to evo's Testing API
As I had anticipated, there are unexpected difficulties when trying to use busted: It can't execute multiple test files in standalone mode and all attempts to work around that have failed.

Since the issues on their repo have been ignored for years, I'm not holding my breath. for them to fix the problem... so back to DIY land we go.