Runtime: Implement lua file loading via the bundle API
There's not much of a distinction between loading a Lua file directly and loading a "folder bundle" (in luvi lingo) which just happens to have a Lua file as its entry point.

The only question is whether or not the -m entry point should override the file passed. I opted to simply ignore the argument if a file is passed and not a directory, as that seems more intuitive than overriding the file loaded...

Example: "Load script.lua with entry point something.lua" would result in loading the latter and not the former, which is pretty weird.