Runtime: Ensure the main thread is wrapped in a coroutine
This is required for the networking tests, and also generally useful. Without it, every script would have to start the event loop itself, when it should be "hidden" and automatic (like in NodeJS) for a better user experience.

This unfortunately makes stack traces less useful since each coroutine has a separate stack, and libuv callbacks also interfere... but that's a problem to figure out later.