Tests: Add smoke tests for code snippets and examples
While it would be ideal to fully test them, as well, I've prototyped some approaches for doing this and determined it isn't worth the hassle. Based on the heuristic that most code changes would lead the examples to fail loudly (raise an error), simply executing them may be enough for the time being.

This decision may have to be revisited if there are more complex examples, or they have side effects. Perhaps it would be better to implement them as acceptance tests (Scenarios)? To be determined...