Tests: Add a test for ignoring SIGPIPE signals in TcpServer
While this doesn't affect Windows, on UNIX the OS can send this signal when a socket has gone away and the server's trying to write to it still. If uncaught, this will kill the runtime, taking the server with it... So that's basically a self-DOS waiting to happen (and it can appear "randomly", so it would be quite annoying to reproduce).

While this test doesn't simulate a real-world scenario in order to keep the complexity low, it does verify that receiving a SIGPIPE won't crash the server, which is good enough. It should be safe to rely on the OS sending the signal if (and only if) it's appropriate, anyway.

This CAN be manually reproduced with the following steps:

1. Start TCP echo server in one shell
2. In another shell: nc localhost 12345 < /dev/zero > /dev/null
3. CTRL/C the netcat shell
4. TCP echo server will receive SIGPIPE while writing to the socket

This proved too annoying to reproduce in code, even with spawning child processes I couldn't get it to work. So I opted to just simulate the signal itself and simply call it a day.