Runtime: Add bindings to the statically linked llhttp API
I don't want to use the standard DLL/SO-based loading for this since it necessitates shipping a separate file with the runtime. I also don't want to write standard Lua bindings because I'm lazy and they're slow in LuaJIT.

Therefore I'm using the approach suggested by the LuaJIT author, which involves exposing the statically-linked APIs via a wrapper object pushed to LuaJIT as lightuserdata.

With this setup, the APIs are exactly the same as when using the DLL-based FFI module, but using some initial setup code the APIs can be added to the llhttp module after the fact and they'll be optimized.

Obviously doesn't work without LuaJIT, but that's a given and not really a problem here.