Refactor: Move the snippet tests to a separate file
Since they may spam the console with irrelevant output, having them run as part of the unit test suite breaks the flow in the console window and makes it more difficult to see at a glance whether any tests have failed.

Of course, it would be possible to just mute the output, but I don't want the examples to run in a modified environment to ensure they' are tested just as they would be running after copy/pasting them straight from the docs.