Refactor: Move the utf8 tests to a separate spec file
Having them as part of the legacy "samples", which currently also double as tests, seems a bit messy. It will be much easier to make sure it works and to add more tests when needed if these apps are converted to "regular" tests, so this is a first step in that direction.