Deps: Add build scripts for MINGW64 on Windows
Invoking each library's native build system from the command line allows debugging the many build failures more easily, and to get rid of CMake for the runtime itself.

Since the dependencies rarely need building, there's not much of an advantage in wiring them into the runtime's build process. In fact, CMake has frequently triggered rebuilds when there was little point in doing so, which is especially agonizing with OpenSSL's lengthy build times... That's just one of the many pain points that should hopefully be eliminated by this more simplistic approach.