Runtime: Add a default value for the LUVI_VERSION define
If the version cannot be obtained from git, there should be a placeholder available that can be displayed in the CLI without errors.

Unit tests can then make sure that the format itself is adhered to, meaning CI failures when the version wasn't properly embedded. Without this, it would also fail but the error message would be less clear ("random" Lua error).