Repo: Convert Windows build scripts to MINGW64 format
Turns out that the MSYS2 shell provided by GitHub runners cannot resolve relative paths correctly when forwarding scripts to CMD - I gave up on debugging that and decided to just use bash scripts.

They work locally and in the runner, and they're also much closer in syntax to the Unix scripts (almost identical, even) which reduces the friction. Let's face it, Windows batch syntax is insane, so that's a nice bonus.

Unfortunately, this means the original convention of having the file extension signal which version it is no longer works. I opted to encode the windowsbuild/unixbuild directly in the file names to resolve this, while also ensuring tab-autocompletion remains relatively painless (you type <libname>-w or -u and that's it).

---

The updated scripts are slightly different to account for some Mac OS shenanigans (failures in the CI), missing OPENSSL_* variables etc. I don't care to disentangle that now, even if it shouldn't be part of this change ideally :(