CI: Update the build workflow to enable git describe usage
The build script fails to discover the tags and so cannot set LUVI_VERSION, which causes the version test to fail (only in the CI).

It should be noted that this has significant performance implications (probably not an issue unless the history becomes extremely large), and there may be race conditions if tags are pushed while a tagged version's CI is running still.

The latter could be problematic when pushing multiple versions shortly after each other, which is unlikely to happen here.