Deps: Add a utility script to export the llhttp cdefs
This just does some minor preprocessing to clear out stuff that LuaJIT can't load via ffi.cde(). I elected to leave this as a manual step because it's annoying to automate reliably for all platforms as part of the build process, so for now whenever llhttp is updated this should also be run.

Not an ideal solution, but enough for the time being. More automation may follow in the future, if the update frequency turns out too high.