Runtime: Remove SetWriteQueueSize from the socket mixin
The "write queue" in libuv is a read-only software-level buffer on top of the actual send buffer. The latter can be accessed via a different API, which apparently I confused with this one when I first added the mixins.

As far as I can tell, the write queue is mainly useful to detect backpressure for socket writes; it will grow whenever the peer stops reading and only afterwards push the data to the underlying send buffer.