Deps: Replace the LPEG mirror with lpeg-compat-52
LPEG defines some 5.2 stuff that clashes with LuaJIT's 5.2 compatibility mode. Since I doubt either of the authors has any interest in fixing that, the next best solution is to just maintain a separate fork with the trivial adaptations needed to resolve the conflict.

I don't expect this to be a huge maintenance burden as LPEG is rarely ever updated.