Runtime: Fix an error in TcpServer.Disconnect
If the client socket is already closing, such as when it disconnected immediately before the server starts shutting down, the libuv bindings will raise an error and the server will crash.

While that may be forgivable when it's already shutting down, it could be problematic under some circumstances.

As for testing this specific scenario: I guess it's covered by the existing acceptance tests for TCP_EOF_RECEIVED already...